pyproject.toml
[tool.poetry]
name = "conversor-moedas-fast-api"
version = "0.1.0"
description = ""
authors = ["fillipeberssot95@hotmail.com <fillipeberssot95@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
requests = "^2.32.3"
aiohttp = "^3.11.12"
pydantic-settings = "^2.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.8.6"
vcrpy = "^7.0.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev conversor_moedas_fast_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=conversor_moedas_fast_api -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
